refactor(): finisih network client and retrofit client
