+++ app/src/main/java/com/ardidong/weatherornot/app/MainActivity.kt
@@ -17,14 +17,14 @@ class MainActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-//        binding = ActivityMainBinding.inflate(layoutInflater)
-//        setContentView(binding.root)
-//
-//        viewModel.observableResult.observe(this){
-//            binding.mainText.text = it.toString()
-//        }
-        setContent {
-            WeatherOrNotApp()
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        viewModel.observableResult.observe(this){
+            binding.mainText.text = it.toString()
         }
+//        setContent {
+//            WeatherOrNotApp()
+//        }
     }
 }
