+++ build.gradle
@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id 'com.android.application' version '7.3.0' apply false
-    id 'com.android.library' version '7.3.0' apply false
+    id 'com.android.application' version '7.4.0' apply false
+    id 'com.android.library' version '7.4.0' apply false
     id 'org.jetbrains.kotlin.android' version '1.7.10' apply false
     id 'com.google.dagger.hilt.android' version '2.44' apply false
 }
@@ -24,6 +24,6 @@ ext{
             hiltKapt                : "com.google.dagger:hilt-compiler:" + hilt_version,
             retrofit                : "com.squareup.retrofit2:retrofit:" + retrofit_version,
             gsonConverter           : "com.squareup.retrofit2:converter-gson:" + retrofit_version,
-            loggingInterceptor       : "com.squareup.okhttp3:logging-interceptor:" + logging_interceptor_version
+            loggingInterceptor      : "com.squareup.okhttp3:logging-interceptor:" + logging_interceptor_version
     ]
 }
