+++ data/library/src/main/java/com/ardidong/weatherornot/library/network/RetrofitClient.kt
@@ -12,31 +12,31 @@ class RetrofitClient @Inject constructor(
     private val retrofit: Retrofit
 ) : NetworkClient {
 
-    override suspend fun <T : Any> getInstance(clazz: KClass<T>): KClass<T> {
-        return retrofit.create(clazz::class.java)
+    override suspend fun <T : Any> getInstance(clazz: Class<T>): T {
+        return retrofit.create(clazz)
     }
 
-}
-
-suspend fun <T : Any> handleApi(
-    execute: suspend () -> Response<T>
-): ResultOf<T> {
-    return try {
-        val response = execute()
-        val body = response.body()
-        if (response.isSuccessful && body != null) {
-            ResultOf.Success(body)
-        } else {
-            ResultOf.Failure(ErrorEntity.ApiResponseError(message = response.message(), errorCode = response.code().toString()))
-        }
-    } catch (e: HttpException) {
-        ResultOf.Failure(
-            ErrorEntity.ApiResponseError(
-                message = e.message(),
-                errorCode = e.code().toString()
+    override suspend fun <T : Any> handleApi(
+        execute: suspend () -> Response<T>
+    ): ResultOf<T> {
+        return try {
+            val response = execute()
+            val body = response.body()
+            if (response.isSuccessful && body != null) {
+                ResultOf.Success(body)
+            } else {
+                ResultOf.Failure(ErrorEntity.ApiResponseError(message = response.message(), errorCode = response.code().toString()))
+            }
+        } catch (e: HttpException) {
+            ResultOf.Failure(
+                ErrorEntity.ApiResponseError(
+                    message = e.message(),
+                    errorCode = e.code().toString()
+                )
             )
-        )
-    } catch (e: Throwable) {
-        ResultOf.Failure(ErrorEntity.ApiResponseError(message = e.message.toString(), errorCode = ""))
+        } catch (e: Throwable) {
+            ResultOf.Failure(ErrorEntity.ApiResponseError(message = e.message.toString(), errorCode = ""))
+        }
     }
+
 }
